Added additional jQuery features
such as animations to user clicks, adding a variable called 'started' to not trigger the nextSequence() function. additionally, inside the nextSequence() function, I updated the h1 to change the value of level. Lastly, the h1 title also starts out saying "Press A Key to Start", when the game has started, because of this, I added a variable called 'level' with the value of 0 so it can start with level 0 once a key has been pressed.